# shared-context/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FinanciAlly API
  version: 0.2.0
  description: |
    Personal finance double-entry ledger. M1 adds the auth surface
    (register/login/refresh/google + protected me/logout). Accounts, entries,
    lines, budgets, fx, recurring, reporting, and sync land in later milestones;
    the Go server interface is regenerated from this file via `make gen`.
servers:
  - url: http://localhost:8080
    description: local dev
...
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      summary: Rotate a refresh token into a new session
      description: |
        The presented refresh token is revoked and a fresh access+refresh pair
        is returned. A replayed, expired, or forged token yields 401; rotation
        is atomic, so reuse is detected on the second presentation.

        `refresh_token` is optional: a browser client posts `{}` and the server
        reads the token from the httpOnly `fa_refresh` cookie instead
        (WEB_COOKIE_AUTH). Mobile keeps sending it in the body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: rotated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: token invalid, expired, or already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: stable machine code for this error class
          example: invalid_credentials
        message:
          type: string
          description: human-readable detail
          example: invalid credentials
    User:
      type: object
      required: [id, email, base_currency, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        base_currency:
          type: string
          description: ISO 4217 currency code
          example: IDR
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
          description: "short-lived HS256 JWT; send as `Authorization: Bearer <token>`"
        refresh_token:
          type: string
          description: single-use opaque token; rotate via /auth/refresh
        user:
          $ref: '#/components/schemas/User'
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: >
            Omit (post `{}`) when the caller is a browser relying on the
            httpOnly fa_refresh cookie. Mobile always sends it.
...